style: format code for consistency and readability in gravity.c

+++ src/gravity.c
@@ -7,82 +7,85 @@
 
 // Define the concrete Simulation struct
 typedef struct {
-    Particle *particles; // Array of particles
-    uint64_t particle_count; // Number of particles
-    bool enable_collisions; // Collision handling flag
-    double time_step; // Time step for the simulation
-    uint32_t substeps; // Number of substeps
+  Particle *particles;     // Array of particles
+  uint64_t particle_count; // Number of particles
+  bool enable_collisions;  // Collision handling flag
+  double time_step;        // Time step for the simulation
+  uint64_t substeps;       // Number of substeps
 } SimulationStruct;
 
 Simulation init_simulation(SimulationOptions options) {
-    // Allocate memory for the simulation
-    SimulationStruct *sim = (SimulationStruct *)malloc(sizeof(SimulationStruct));
-    if (!sim) {
-        return NULL; // Allocation failed
-    }
+  // Allocate memory for the simulation
+  SimulationStruct *sim = (SimulationStruct *)malloc(sizeof(SimulationStruct));
+  if (!sim) {
+    return NULL; // Allocation failed
+  }
 
-    // Initialize the simulation with the given options
-    sim->particle_count = options.particle_count;
-    sim->enable_collisions = options.enable_collisions;
-    sim->time_step = options.time_step;
-    sim->substeps = options.substeps;
+  // Initialize the simulation with the given options
+  sim->particle_count = options.particle_count;
+  sim->enable_collisions = options.enable_collisions;
+  sim->time_step = options.time_step;
+  sim->substeps = options.substeps;
 
-    // Allocate memory for particles
-    sim->particles = (Particle *)malloc(sizeof(Particle) * sim->particle_count);
-    if (!sim->particles) {
-        free(sim);
-        return NULL; // Allocation failed
-    }
+  // Allocate memory for particles
+  sim->particles = (Particle *)malloc(sizeof(Particle) * sim->particle_count);
+  if (!sim->particles) {
+    free(sim);
+    return NULL; // Allocation failed
+  }
 
-    // Initialize particles (this is a placeholder, actual initialization may vary)
-    for (uint64_t i = 0; i < sim->particle_count; ++i) {
-        sim->particles[i] = (Particle){0}; // Initialize each particle
-    }
+  // Initialize particles (this is a placeholder, actual initialization may
+  // vary)
+  for (uint64_t i = 0; i < sim->particle_count; ++i) {
+    sim->particles[i] = (Particle){0}; // Initialize each particle
+  }
 
-    return (Simulation)sim;
+  return (Simulation)sim;
 }
 
 void deinit_simulation(Simulation sim) {
-    SimulationStruct *sim_struct = (SimulationStruct *)sim;
-    if (sim_struct) {
-        // Free the particles array
-        free(sim_struct->particles);
-        // Free the simulation struct
-        free(sim_struct);
-    }
+  SimulationStruct *sim_struct = (SimulationStruct *)sim;
+  if (sim_struct) {
+    // Free the particles array
+    free(sim_struct->particles);
+    // Free the simulation struct
+    free(sim_struct);
+  }
 }
 
 void step_simulation(Simulation sim) {
-    // Update the simulation by the given time step
+  // Update the simulation by the given time step
 }
 
 void toggle_collisions(bool enable) {
-    // Enable or disable collision handling
+  // Enable or disable collision handling
 }
 
 Particle get_particle_state(int particle_id) {
-    // Return the current state of the particle
-    Particle particle; // Placeholder return
-    return particle;
+  // Return the current state of the particle
+  Particle particle; // Placeholder return
+  return particle;
 }
 
-int get_particles_by_area(Vector2D top_left, Vector2D bottom_right, ParticleInfo* buffer, int max_count) {
-    // Retrieve particles within the specified area
-    return 0; // Placeholder return
+int get_particles_by_area(Vector2D top_left, Vector2D bottom_right,
+                          ParticleInfo *buffer, int max_count) {
+  // Retrieve particles within the specified area
+  return 0; // Placeholder return
 }
 
 // Adds a new particle to the simulation.
-void add_particle(double mass, double size, Vector2D position, Vector2D velocity) {
-    // Add a new particle with specified properties
+void add_particle(double mass, double size, Vector2D position,
+                  Vector2D velocity) {
+  // Add a new particle with specified properties
 }
 
 // Removes a particle from the simulation.
 void remove_particle(int particle_id) {
-    // Remove the particle with the given ID
+  // Remove the particle with the given ID
 }
 
 // Applies an external force to a specific particle.
 void apply_force(int particle_id, Vector2D force) {
-    // Apply the specified force to the particle
+  // Apply the specified force to the particle
 }
 #endif // GRAVITY_C
